adding config.h
git-svn-id: svn+ssh://hamsterdb.com/home/chris/repos/hamsterdb/trunk@1262 66673985-dd14-0410-9433-caba4d15c716

+++ unittests/log.cpp
@@ -9,6 +9,8 @@
  * See files COPYING.* for License information.
  */
 
+#include "../src/config.h"
+
 #include <stdexcept>
 #include <cstring>
 #include <vector>
@@ -896,7 +898,7 @@ public:
                 void *pl=&(*itl).m_data[0];
                 void *pr=&(*itr).m_data[0];
                 BFC_ASSERT_EQUAL(0, memcmp(pl, pr, 
-                            log_entry_get_data_size(&(*itl).m_entry)));
+                            (size_t)log_entry_get_data_size(&(*itl).m_entry)));
             }
         }
     }
